# postman/specs/Legal API/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Legal API",
    "version": "1.0.0",
    "description": "API documentation for Legal backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ObjectId": {
        "type": "string",
        "example": "681d0f8f8d7c4b0012f34abc"
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 10
          }
        }
      },
...
      "CaseStatus": {
        "type": "string",
        "enum": [
          "active",
          "pending",
          "completed",
          "cancelled",
          "closed"
        ],
        "example": "active"
      },
...
      "PaymentType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ],
        "example": "income"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "completed",
          "cancelled"
        ],
        "example": "completed"
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "cash",
          "transfer",
          "card",
          "check",
          "other"
        ],
        "example": "transfer"
      },
...
      "ActivityAction": {
        "type": "string",
        "enum": [
          "create_case",
          "update_case",
          "delete_case",
          "create_stage",
          "update_stage",
          "complete_stage",
          "upload_document",
          "create_note",
          "register_payment",
          "login",
          "logout"
        ],
        "example": "create_case"
      },
...
      "FinancialSummary": {
        "type": "object",
        "properties": {
          "totalCost": {
            "type": "number",
            "example": 5000
          },
          "totalPaid": {
            "type": "number",
            "example": 2500
          },
          "pendingAmount": {
            "type": "number",
            "example": 2500
          },
          "expenses": {
            "type": "number",
            "example": 1200
          }
        }
      },
      "Case": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "code": {
            "type": "string",
            "example": "CASE-2025-001"
          },
          "title": {
            "type": "string",
            "example": "Proceso laboral"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true,
            "example": "laboral"
          },
          "clientId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "createdBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "principalLawyerId": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "assignedUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "example": "high"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "estimatedEndDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "currentStageId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "urgente",
              "laboral"
            ]
          },
          "financialSummary": {
            "$ref": "#/components/schemas/FinancialSummary"
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateCaseDTO": {
        "type": "object",
        "required": [
          "code",
          "title",
          "clientId",
          "principalLawyerId",
          "createdBy"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "CASE-2025-001"
          },
          "title": {
            "type": "string",
            "example": "Proceso laboral"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "example": "laboral"
          },
          "clientId": {
            "type": "string"
          },
          "principalLawyerId": {
            "type": "string"
          },
          "assignedUsers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "priority": {
            "type": "string",
            "example": "high"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "estimatedEndDate": {
            "type": "string",
            "format": "date-time"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
      "UpdateCaseDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "principalLawyerId": {
            "type": "string"
          },
          "assignedUsers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "priority": {
            "type": "string"
          },
          "estimatedEndDate": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "currentStageId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "caseId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "uploadedBy": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "stageId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "name": {
            "type": "string",
            "example": "Contrato firmado"
          },
          "originalName": {
            "type": "string",
            "example": "contrato.pdf"
          },
          "mimeType": {
            "type": "string",
            "example": "application/pdf"
          },
          "size": {
            "type": "number",
            "example": 1500000
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "visibility": {
            "type": "string",
            "example": "internal"
          },
          "documentType": {
            "type": "string",
            "nullable": true,
            "example": "contract"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "contrato",
              "firmado"
            ]
          },
          "version": {
            "type": "number",
            "example": 1
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Activity": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "caseId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "stageId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "action": {
            "$ref": "#/components/schemas/ActivityAction"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "El usuario creó un nuevo caso."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CasePayment": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "caseId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "amount": {
            "type": "number",
            "example": 1500
          },
          "type": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "category": {
            "type": "string",
            "example": "Honorarios"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true,
            "example": "TRX-001"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "registeredBy": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "updatedBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "deletedBy": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "affectsGlobalFinance": {
            "type": "boolean",
            "example": true
          },
          "financeTransactionId": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ObjectId"
              }
            ]
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CasePaymentListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CasePayment"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/activities/case/{caseId}": {
      "get": {
        "summary": "Obtener actividades de un caso",
        "tags": [
          "Activities"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "responses": {
          "200": {
            "description": "Actividades del caso obtenidas correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Activity"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      }
    },
...
    "/cases": {
      "post": {
        "summary": "Crear un nuevo caso",
        "tags": [
          "Cases"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Caso creado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Case"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "get": {
        "summary": "Obtener todos los casos",
        "tags": [
          "Cases"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por estado"
          },
          {
            "in": "query",
            "name": "clientId",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por cliente"
          },
          {
            "in": "query",
            "name": "principalLawyerId",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por abogado principal"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por prioridad"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Búsqueda general"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Cantidad de resultados"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de casos obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Case"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/cases/{id}": {
      "get": {
        "summary": "Obtener un caso por ID",
        "tags": [
          "Cases"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "responses": {
          "200": {
            "description": "Caso obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Case"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un caso",
        "tags": [
          "Cases"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCaseDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Caso actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Case"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un caso",
        "tags": [
          "Cases"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "responses": {
          "200": {
            "description": "Caso eliminado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Caso eliminado correctamente"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      }
    },
...
    "/clients/{id}/cases": {
      "get": {
        "summary": "Obtener casos de un cliente",
        "tags": [
          "Clients"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del cliente"
          }
        ],
        "responses": {
          "200": {
            "description": "Casos del cliente obtenidos correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Case"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        }
      }
    },
...
    "/documents/case/{caseId}": {
      "get": {
        "summary": "Obtener documentos de un caso",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de documentos obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      }
    },
...
    "/case-payments/case/{caseId}": {
      "get": {
        "summary": "Obtener pagos de un caso",
        "tags": [
          "Case Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "caseId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del caso"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pagos obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePaymentListResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Caso no encontrado"
          }
        }
      }
    },
...
  },
...
}